requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
